community-ing8

+++ front-end/src/views/community/CommunityDetailView.vue
@@ -168,26 +168,40 @@ export default {
 
 
         deleteArticle(){
-            const axiosConfig2 = {
-              headers:{
-                token: `${this.$cookies.get('auth-token')}`,
+            this.$confirm(
+                {
+                message: `삭제하시겠습니까?`,
+                button: {
+                    yes: '삭제하기',
+                    no: '아니요',
                 },
-            }
-            axios.delete(SERVER.URL+`/community/delete/${this.communityIdx}`,axiosConfig2)
-            .then((response)=>{
-                console.log(response)
-                this.$alert('삭제 완료')
-                this.$router.push({name:'CommunityIndexView'})                
-            })
-            .catch((err)=>{
-                console.log(err)
-            })
+                callback: confirm => {
+                    if (confirm) {
+                        const axiosConfig2 = {
+                        headers:{
+                            token: `${this.$cookies.get('auth-token')}`,
+                            },
+                        }
+                        axios.delete(SERVER.URL+`/community/delete/${this.communityIdx}`,axiosConfig2)
+                        .then((response)=>{
+                            console.log(response)
+                            this.$alert('삭제 완료')
+                            this.$router.push({name:'CommunityIndexView'})                
+                        })
+                        .catch((err)=>{
+                            console.log(err)
+                        })                  
+                    }
+                }
+                }
+            )
         },
 
 
         editArticle(){
             this.$router.push({ name: 'CommunityUpdateView', params: { communityId: this.communityIdx, page:this.$route.params.page }})
         },
+
 
         addcomment(){
             this.commentCnt += 1
